pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copilot-pulse"
version = "0.1.0"
description = "AI-powered conversational agent for GitHub Copilot usage analytics"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "httpx>=0.27,<1",
    "pydantic>=2.0,<3",
    "rich>=13.0,<14",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "jinja2>=3.1,<4",
    "plotly>=5.20,<6",
    "matplotlib>=3.9,<4",
    "openpyxl>=3.1,<4",
    "reportlab>=4.1,<5",
    "python-dotenv>=1.0,<2",
    "anthropic>=0.40,<1",
    "click>=8.1,<9",
    "websockets>=13.0,<14",
    "kaleido==0.2.1",
    "pyjwt[crypto]>=2.8,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-httpx>=0.30,<1",
    "ruff>=0.5,<1",
]

[project.scripts]
copilot-pulse = "src.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
